Unit06/pampered_pets_app/requirements.txt
matplotlib==3.7.1
numpy==1.24.2
pyyaml==6.0
lxml==4.9.2
